# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API",
    "version": "1.0.0",
    "description": "API documentation for authentication endpoints"
  },
  "servers": [
    {
      "url": "https://api-mkcy.onrender.com"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "email": { "type": "string", "format": "email" }
                },
                "required": ["username", "password", "email"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully and welcome email sent"
          },
          "400": {
            "description": "Missing or invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing input"
          },
          "401": {
            "description": "Invalid username or password"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Send OTP to user's email for password reset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent to email"
          },
          "400": {
            "description": "Email is required"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "summary": "Change password using OTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "otp": { "type": "string" },
                  "newPassword": { "type": "string" }
                },
                "required": ["email", "otp", "newPassword"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid OTP or OTP expired"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}